+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+### [1.5.3]
+
+-   Add `onFeaturesLoadStart` and `onFeaturesLoadError` events to all vector layers
+
 ### [1.5.2] 2023-06-11
 
 -   Support OpenLayers 7.4.0